Move inference code to library dir (#801)
* move inference engine

* move file functions

* move schema

* minor fix

* minor fix

* move inference code

* update docs

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* minor fix

* [feature]retain the interface to support old version codes

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update inference.ipynb tp support 1.5

* Remove unused package

---------
