+++ fish_speech/inference_engine/__init__.py
@@ -6,18 +6,18 @@ import numpy as np
 import torch
 from loguru import logger
 
-from fish_speech.models.vqgan.modules.firefly import FireflyArchitecture
-from fish_speech.text.chn_text_norm.text import Text as ChnNormedText
-from fish_speech.utils import autocast_exclude_mps, set_seed
-from tools.inference_engine.reference_loader import ReferenceLoader
-from tools.inference_engine.utils import InferenceResult, wav_chunk_header
-from tools.inference_engine.vq_manager import VQManager
-from tools.llama.generate import (
+from fish_speech.inference_engine.reference_loader import ReferenceLoader
+from fish_speech.inference_engine.utils import InferenceResult, wav_chunk_header
+from fish_speech.inference_engine.vq_manager import VQManager
+from fish_speech.models.text2semantic.inference import (
     GenerateRequest,
     GenerateResponse,
     WrappedGenerateResponse,
 )
-from tools.schema import ServeTTSRequest
+from fish_speech.models.vqgan.modules.firefly import FireflyArchitecture
+from fish_speech.text.chn_text_norm.text import Text as ChnNormedText
+from fish_speech.utils import autocast_exclude_mps, set_seed
+from fish_speech.utils.schema import ServeTTSRequest
 
 
 class TTSInferenceEngine(ReferenceLoader, VQManager):
@@ -72,7 +72,10 @@ class TTSInferenceEngine(ReferenceLoader, VQManager):
         if req.streaming:
             yield InferenceResult(
                 code="header",
-                audio=(sample_rate, wav_chunk_header(sample_rate=sample_rate)),
+                audio=(
+                    sample_rate,
+                    np.array(wav_chunk_header(sample_rate=sample_rate)),
+                ),
                 error=None,
             )
