--- fish_speech/webui/html/footer.html
@@ -1,11 +1,0 @@
-<div style="color: rgba(25,255,205,0.7) !important;">
-        <a href="{api_docs}">API</a>
-         • 
-        <a href="https://github.com/fishaudio/fish-speech">Github</a>
-         • 
-        <a href="https://gradio.app">Gradio</a>
-</div>
-<br />
-<div class="versions" style="color: rgba(25,255,205,0.7) !important;">
-{versions}
-</div>
