+++ tools/llama/build_dataset.py
@@ -12,8 +12,7 @@ from loguru import logger
 from tqdm import tqdm
 
 from fish_speech.datasets.protos.text_data_pb2 import Semantics, Sentence, TextData
-from fish_speech.datasets.protos.text_data_stream import pack_pb_stream
-from tools.file import load_filelist
+from fish_speech.datasets.protos.text_data_stream import load_filelist, pack_pb_stream
 
 # To avoid CPU overload
 os.environ["MKL_NUM_THREADS"] = "1"
