+++ tools/run_webui.py
@@ -8,10 +8,10 @@ from loguru import logger
 
 pyrootutils.setup_root(__file__, indicator=".project-root", pythonpath=True)
 
-from tools.inference_engine import TTSInferenceEngine
-from tools.llama.generate import launch_thread_safe_queue
-from tools.schema import ServeTTSRequest
-from tools.vqgan.inference import load_model as load_decoder_model
+from fish_speech.inference_engine import TTSInferenceEngine
+from fish_speech.models.text2semantic.inference import launch_thread_safe_queue
+from fish_speech.models.vqgan.inference import load_model as load_decoder_model
+from fish_speech.utils.schema import ServeTTSRequest
 from tools.webui import build_app
 from tools.webui.inference import get_inference_wrapper
